mvp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.7.0
openai==1.40.0
tavily-python==0.3.9
stripe==10.5.0
pydantic==2.8.0
httpx==0.27.0
